Move scrollbar inside the navigation container.
